15.09.2022

+++ app/src/main/java/uz/jbnuu/tsc/model/semester/WeekData.kt
@@ -4,8 +4,8 @@ data class WeekData(
     val id: Int?,
     val _semester: String?,
     val current: Boolean?,
-    val start_date: Int?,
-    val end_date: Int?,
+    val start_date: Long?,
+    val end_date: Long?,
     val start_date_f: String?,
     val end_date_f: String?,
 )
